Add files via upload
Determine if a truck's tag ("DDXDDD-DD") is valid by ensuring the letter is not a vowel and the sum of every two consecutive digits at specific positions is even.

We consider the letters "A","E","I","O","U","Y" to be vowels for this problem.

Input Format

The first line contains a string of length 9. The format is "DDXDDD-DD", where D stands for a digit (non zero) and X is an uppercase english letter.